+++ private-templates-service/client/src/components/Common/Tags/index.tsx
@@ -1,0 +1,3 @@
+import Tags from './Tags';
+
+export default Tags;
